+++ src/Completion/CompletionSource.cs
@@ -1,11 +1,10 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using Microsoft.VisualStudio.Imaging;
-using Microsoft.VisualStudio.Imaging.Interop;
+﻿using Microsoft.VisualStudio.Imaging;
 using Microsoft.VisualStudio.Language.Intellisense;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Classification;
 using Microsoft.VisualStudio.Text.Operations;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace EditorConfig
 {
@@ -42,7 +41,7 @@ namespace EditorConfig
             var applicableTo = snapshot.CreateTrackingSpan(position, 0, SpanTrackingMode.EdgeInclusive);
 
             var parseItem = _document.ItemAtPosition(position);
-            var prev = _document.ParseItems.LastOrDefault(p => p.Span.Start < position && !p.Span.Contains(position));
+            var prev = _document.ParseItems.LastOrDefault(p => p.Span.Start < position && !p.Span.Contains(position - 1));
 
             // Property
             if (string.IsNullOrWhiteSpace(line.GetText()) || parseItem?.ItemType == ItemType.Property)
@@ -67,15 +66,15 @@ namespace EditorConfig
             // Severity
             else if ((position > 0 && snapshot.Length > 1 && snapshot.GetText(position - 1, 1) == ":") || parseItem?.ItemType == ItemType.Severity)
             {
-                if (parseItem?.ItemType == ItemType.Severity)
-                {
-                    AddSeverity(list);
-                }
-                else
-                {
+                //if (parseItem?.ItemType == ItemType.Severity)
+                //{
+                //    AddSeverity(list);
+                //}
+                //else
+                //{
                     if (SchemaCatalog.TryGetProperty(prev?.Prev.Text, out Keyword prop) && prop.SupportsSeverity)
                         AddSeverity(list);
-                }
+                //}
             }
 
             if (!list.Any())