+++ src/Parser/EditorConfigDocumentValidator.cs
@@ -40,31 +40,33 @@ namespace EditorConfig
 
         private void Validate()
         {
-            foreach (var item in ParseItems)
+            try
             {
-                switch (item.ItemType)
+                foreach (var item in ParseItems.Where(i => i.ItemType == ItemType.Unknown))
                 {
-                    case ItemType.Unknown:
-                        ValidateUnknown(item);
-                        break;
+                    ValidateUnknown(item);
                 }
+
+                ValidateSection();
+
+                foreach (var property in Properties)
+                {
+                    ValidateProperty(property);
+                }
+
+                Validated?.Invoke(this, EventArgs.Empty);
             }
-
-            ValidateSection();
-
-            foreach (var property in Properties)
+            catch (Exception ex)
             {
-                ValidateProperty(property);
+                System.Diagnostics.Debug.Write(ex);
             }
-
-            Validated?.Invoke(this, EventArgs.Empty);
         }
 
         private void ValidateUnknown(ParseItem item)
         {
             item.AddError("Syntax error. Element not valid at current location");
         }
-        
+
         private void ValidateSection()
         {
             foreach (var section in Sections)
@@ -72,11 +74,13 @@ namespace EditorConfig
                 foreach (var property in section.Properties)
                 {
                     ValidateProperty(property);
-
+                    
+                    // Duplicate property
                     if (section.Properties.First(p => p.Keyword.Text.Equals(property.Keyword.Text, StringComparison.OrdinalIgnoreCase)) != property)
                         property.Keyword.AddError(Resources.Text.ValidationDuplicateProperty);
                 }
 
+                // Duplicate section
                 if (Sections.First(s => s.Item.Text == section.Item.Text) != section)
                     section.Item.AddError(string.Format(Resources.Text.ValidationDuplicateSection, section.Item.Text));
             }
@@ -86,6 +90,7 @@ namespace EditorConfig
         {
             foreach (var property in Properties)
             {
+                // Only root property allowed
                 if (property != Root)
                     property.Keyword.AddError(Resources.Text.ValidationRootInSection);
             }
@@ -99,24 +104,31 @@ namespace EditorConfig
                 property.Keyword.AddError(string.Format(Resources.Text.ValidateUnknownKeyword, property.Keyword.Text));
             }
 
-            // Value
-            else if (!keyword.Values.Any(v => v.Name.Equals(property.Value.Text, StringComparison.OrdinalIgnoreCase)) &&
+            // Missing value
+            else if (property.Value == null)
+            {
+                property.Keyword.AddError("A value must be specified");
+            }
+            // Value not in schema
+            else if (!keyword.Values.Any(v => v.Name.Equals(property.Value?.Text, StringComparison.OrdinalIgnoreCase)) &&
                 !(int.TryParse(property.Value.Text, out int intValue) && intValue > 0))
             {
                 property.Value.AddError(string.Format(Resources.Text.InvalidValue, property.Value.Text, keyword.Name));
             }
-
-            // Severity
+            
+            // Missing severity
             else if (property.Severity == null && property.Value.Text.Equals("true", StringComparison.OrdinalIgnoreCase) && keyword.SupportsSeverity)
             {
                 property.Value.AddError(Resources.Text.ValidationMissingSeverity);
             }
             else if (property.Severity != null)
             {
+                // Severity not applicaple to property
                 if (!keyword.SupportsSeverity)
                 {
                     property.Severity.AddError(string.Format("The \"{0}\" property does not support a severity suffix", keyword.Name));
                 }
+                // Severity not in schema
                 else if (!SchemaCatalog.TryGetSeverity(property.Severity.Text, out Severity severity))
                 {
                     property.Severity.AddError(string.Format(Resources.Text.ValidationInvalidSeverity, property.Severity.Text));